backend/requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangGraph and LangChain
langgraph==0.0.20
langchain==0.1.0
langchain-core>=1.2.1
langchain-groq>=0.1.0

# Groq API
groq>=0.9.0

# HTTP requests
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# JSON handling
orjson==3.9.10

# Streamlit (Demo UI)
streamlit>=1.30.0
